Relocate plan progress helpers to actor_loop_flow
Move 4 plan-related helpers (summarize_plan_write,
progress_stage_label, log_plan_stall, normalize_plan_exploration_key)
out of turn.rs into the sibling actor_loop_flow module. Their callers
already lived in actor_loop_flow, so the relocation removes 3
super::turn:: prefixes and one of the lingering back-imports while
keeping turn.rs's two production call sites + tests routed through the
new home via super::actor_loop_flow::.

Visibility upgrades (private -> pub(super)) for the still-in-turn.rs
dependencies the moved helpers transitively touch:
plan_section_excerpt, plan_write_status, plan_phase_from_sections,
plan_path_matches, normalize_exploration_path,
plan_write_previous_contents, build_plan_write_section_delta,
join_sections_for_progress, and the PlanWriteSectionDelta struct +
fields.

turn.rs LOC: 35,612 -> 35,452 (-160), continuing Phase 1 progress
toward the <=35,000 acceptance criterion (Issue #681 / meta #680).
No production-binary behaviour change, no facade re-export
(DR3-001 maintained).
